backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.11.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
python-multipart==0.0.20
google-generativeai==0.8.5
edge-tts>=6.1.0
mutagen==1.47.0
moviepy==2.1.2
ffmpeg-python==0.2.0
Pillow>=9.2.0,<11.0
resend>=2.0.0
python-telegram-bot>=21.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
